Use TX data for echo compensation as given refered RX data
The hardware driver knows the delay, so it knows which TX data refers to RX
data. The 'id' attribute of RX data message is used to indicate the echo
canceler's TX data ring buffer offset.

For each received RX data, the echo canceler uses the given offset to select
the TX data that refers to the echo.

For echo compensation it is required to have absolutely no shift (jitter)
between TX and RX data, so the hardware driver must care about correct
offset.

Description is bad and is TDB...